.github/ACDbot/pyproject.toml
[build-system]
requires = ["setuptools>=66.1.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ACDbot"
version = "1.0.0"
description = "A bot for handling issues."
authors = [
    { name = "Nicolas Consigny", email = "nicolas.consigny@ethereum.org" }
]
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["github-actions", "automation", "zoom", "discourse"]

# Dependencies
dependencies = [
    "requests>=2.25.1",
    "python-dateutil>=2.8.1",
    "google-api-python-client>=2.0.0",
    "google-auth>=1.30.0",
    "click>=7.1.2",
    "requests-oauthlib",
    "PyGithub>=1.55.1",
    "python-dotenv>=0.20.0"
]
